backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF Processing
pdfplumber==0.10.3

# AI/ML
langchain==0.1.4
langchain-openai==0.0.5
langchain-google-genai==0.0.6
langchain-community==0.0.16
chromadb==0.4.22
openai==1.10.0
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
